backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv
google-genai
pillow
pytesseract
opencv-python
